Add tokenizer test harness
This commit adds a `snap.zig` file which contains a `SnapshotTestSession`
struct. This struct is responsible for reading test snapshots from the
`snap/` folder one by one, based on the file extension. The caller can
then do whatever it wants with the test input and submit its result back
to the session, which will write it to `test-out` and assert the result
equals some expected output file.